Change borders in footer
I updated to tailwind latest to be able to use the arbitrary border values.
