+++ package.json
@@ -47,15 +47,15 @@
     "use-persisted-state": "^0.3.3"
   },
   "devDependencies": {
-    "autoprefixer": "^10.2.6",
+    "autoprefixer": "^10.4.13",
     "concurrently": "^7.6.0",
     "env-cmd": "^10.1.0",
     "eslint": "8.4.1",
     "eslint-config-next": "^13.1.1",
     "eslint-config-prettier": "^8.5.0",
     "husky": "^8.0.2",
-    "postcss": "^8.3.5",
+    "postcss": "^8.4.21",
     "prettier": "2.5.1",
-    "tailwindcss": "^2.2.4"
+    "tailwindcss": "^3.2.7"
   }
 }
