+++ tailwind.config.js
@@ -1,3 +1,5 @@
+
+
 module.exports = {
   mode: 'jit',
   purge: [
@@ -7,7 +9,6 @@ module.exports = {
     './lib/**/*.{js,ts,jsx,tsx}',
     './utils/**/*.{js,ts,jsx,tsx}',
   ],
-  darkMode: false,
   theme: {
     extend: {
       colors: {
